+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 * IBAN country / address country default is Germany "DE" #APPS-589
+* Reset IBAN number on SEPA payment remove action #APPS-590
+* added IBANFormatter for IBAN number textfield #APPS-593
 
 ## [0.25.0] - 2022-12-14